Added protecting resources with authorization

+++ backend/routes/posts.js
@@ -92,11 +92,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message : "Post deleted!"
-    });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: "Deletion Succesfully"});
+    }
+    else {
+      res.status(401).json({message: "Not authorized"});
+    }
   });
 
 });
@@ -118,9 +120,14 @@ multer({storage: storage}).single("image"),
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Updated Succesfully"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: "Updated Succesfully"});
+    }
+    else {
+      res.status(401).json({message: "Not authorized"});
+    }
+
   });
 });
